feat: implementacion subida de archivos a drive en .net completado

+++ backend/FileManagement/FileManagement.Core/Entities/File.cs
@@ -6,7 +6,7 @@ namespace FileManagement.Core.Entities
     {
         public string FileName { get; set; }
         public string Extension { get; set; }
-        public int SizeBytes { get; set; }
+        public long SizeBytes { get; set; }
         public int FolderId { get; set; }
         public Folder Folder { get; set; }
         public FileStorage FileStorage { get; set; }
